refactor: move some types to clientUtils
DimensionProperty, ColumnSlug. Cascading from needing LegacyChartDimensionInterface in model/Variable.ts
