+++ adminSiteClient/VariableEditPage.tsx
@@ -21,12 +21,9 @@ import { GrapherFigureView } from "../site/GrapherFigureView"
 import { ChartList, ChartListItem } from "./ChartList"
 import { AdminAppContext, AdminAppContextType } from "./AdminAppContext"
 import { Base64 } from "js-base64"
-import {
-    DimensionProperty,
-    GrapherTabOption,
-} from "../grapher/core/GrapherConstants"
+import { GrapherTabOption } from "../grapher/core/GrapherConstants"
 import { GrapherInterface } from "../grapher/core/GrapherInterface"
-import { EPOCH_DATE } from "../clientUtils/owidTypes"
+import { DimensionProperty, EPOCH_DATE } from "../clientUtils/owidTypes"
 
 interface VariablePageData extends Omit<LegacyVariableConfig, "source"> {
     datasetNamespace: string
