+++ clientUtils/owidTypes.ts
@@ -248,3 +248,14 @@ export interface Annotation {
     entityName?: string
     year?: number
 }
+
+export enum DimensionProperty {
+    y = "y",
+    x = "x",
+    size = "size",
+    color = "color",
+    table = "table",
+}
+
+// see CoreTableConstants.ts
+export type ColumnSlug = string // a url friendly name for a column in a table. cannot have spaces
