+++ coreTable/CoreTable.ts
@@ -15,7 +15,6 @@ import {
 import { isPresent } from "../clientUtils/isPresent"
 import { CoreColumn, ColumnTypeMap, MissingColumn } from "./CoreTableColumns"
 import {
-    ColumnSlug,
     CoreColumnStore,
     CoreRow,
     CoreTableInputOption,
@@ -66,6 +65,7 @@ import {
 } from "./ErrorValues"
 import { OwidTableSlugs } from "./OwidTableConstants"
 import { applyTransforms } from "./Transforms"
+import { ColumnSlug } from "../clientUtils/owidTypes"
 
 interface AdvancedOptions {
     tableDescription?: string
