+++ coreTable/OwidTableSynthesizers.ts
@@ -4,11 +4,12 @@ import {
     range,
 } from "../clientUtils/Util"
 import { countries } from "../clientUtils/countries"
-import { ColumnSlug, TimeRange } from "./CoreTableConstants"
+import { TimeRange } from "./CoreTableConstants"
 import { ColumnTypeNames } from "./CoreColumnDef"
 import { LegacyVariableDisplayConfigInterface } from "../clientUtils/LegacyVariableDisplayConfigInterface"
 import { OwidTable } from "./OwidTable"
 import { OwidColumnDef, OwidTableSlugs } from "./OwidTableConstants"
+import { ColumnSlug } from "../clientUtils/owidTypes"
 
 interface SynthOptions {
     entityCount: number
