+++ grapher/chart/ChartManager.ts
@@ -9,12 +9,11 @@ import {
 import { ComparisonLineConfig } from "../scatterCharts/ComparisonLine"
 import { TooltipProps } from "../tooltip/TooltipProps"
 import { OwidTable } from "../../coreTable/OwidTable"
-import { ColumnSlug } from "../../coreTable/CoreTableConstants"
 import { AxisConfigInterface } from "../axis/AxisConfigInterface"
 import { ColorSchemeName } from "../color/ColorConstants"
 import { EntityName } from "../../coreTable/OwidTableConstants"
 import { SelectionArray } from "../selection/SelectionArray"
-import { Annotation } from "../../clientUtils/owidTypes"
+import { Annotation, ColumnSlug } from "../../clientUtils/owidTypes"
 
 // The possible options common across our chart types. Not all of these apply to every chart type, so there is room to create a better type hierarchy.
