+++ grapher/dataTable/DataTable.sample.ts
@@ -1,5 +1,6 @@
+import { DimensionProperty } from "../../clientUtils/owidTypes"
 import { Grapher } from "../core/Grapher"
-import { DimensionProperty, GrapherTabOption } from "../core/GrapherConstants"
+import { GrapherTabOption } from "../core/GrapherConstants"
 import { GrapherInterface } from "../core/GrapherInterface"
 
 export const childMortalityGrapher = (
