+++ grapher/dataTable/DataTable.tsx
@@ -4,7 +4,7 @@ import { observer } from "mobx-react"
 import classnames from "classnames"
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
 import { faInfoCircle } from "@fortawesome/free-solid-svg-icons/faInfoCircle"
-import { SortOrder, ColumnSlug, Time } from "../../coreTable/CoreTableConstants"
+import { SortOrder, Time } from "../../coreTable/CoreTableConstants"
 import { EntityName, OwidTableSlugs } from "../../coreTable/OwidTableConstants"
 import { TickFormattingOptions } from "../../clientUtils/formatValue"
 import {
@@ -27,6 +27,7 @@ import { CoreColumn } from "../../coreTable/CoreTableColumns"
 import { Bounds, DEFAULT_BOUNDS } from "../../clientUtils/Bounds"
 import { makeSelectionArray } from "../chart/ChartUtils"
 import { SelectionArray } from "../selection/SelectionArray"
+import { ColumnSlug } from "../../clientUtils/owidTypes"
 
 interface DataTableState {
     sort: DataTableSortState
