+++ grapher/mapCharts/MapChartConstants.ts
@@ -4,9 +4,10 @@ import { PointVector } from "../../clientUtils/PointVector"
 import { MapProjectionName } from "./MapProjections"
 import { ChartManager } from "../chart/ChartManager"
 import { MapConfig } from "./MapConfig"
-import { Color, ColumnSlug, Time } from "../../coreTable/CoreTableConstants"
+import { Color, Time } from "../../coreTable/CoreTableConstants"
 import { ChartTypeName, SeriesName } from "../core/GrapherConstants"
 import { ChartSeries } from "../chart/ChartInterface"
+import { ColumnSlug } from "../../clientUtils/owidTypes"
 
 export type GeoFeature = GeoJSON.Feature<GeoJSON.GeometryObject>
 export type MapBracket = ColorScaleBin
