+++ site/GrapherPage.jsdom.test.tsx
@@ -4,10 +4,13 @@ import * as React from "react"
 import { GrapherPage } from "./GrapherPage"
 import { SiteHeader } from "./SiteHeader"
 import { SiteFooter } from "./SiteFooter"
-import { PostRow, RelatedChart } from "../clientUtils/owidTypes"
+import {
+    DimensionProperty,
+    PostRow,
+    RelatedChart,
+} from "../clientUtils/owidTypes"
 import { ChartListItemVariant } from "./ChartListItemVariant"
 import { GrapherInterface } from "../grapher/core/GrapherInterface"
-import { DimensionProperty } from "../grapher/core/GrapherConstants"
 
 import { configure, shallow, ShallowWrapper } from "enzyme"
 import Adapter from "enzyme-adapter-react-16"
